Refactor as_dataset as a function.
